--- apps/web/src/entities/auction/constants/auctionForm.ts
@@ -1,9 +1,0 @@
-//NOTE - auctionForm의 zod 유효성 검사에 사용하는 상수 목록
-export const MIN_TITLE_LETTERS = 5; //NOTE - 제목의 최소 글자 수
-export const MAX_TITLE_LETTERS = 50; //NOTE - 제목의 최대 글자 수
-export const MIN_DESCRIPTION_LETTERS = 5; //NOTE - 상세 내용의 최소 글자 수
-export const MAX_DESCRIPTION_LETTERS = 500; //NOTE - 상세 내용의 최대 글자 수
-export const MIN_END_TIME_LETTERS = 1; //NOTE - 경매 종료 시각 입력 여부
-export const MIN_STARTING_POINT_NUM = 0; //NOTE - 최소 경매 시작가
-export const MIN_MAX_POINT_NUM = 0; //NOTE - 최대 경매 시작가
-export const BUCKET_FOLDER_NAME = 'images/'; //NOTE - 버켓 폴더 이름
