--- apps/web/src/entities/auction/constants/queryKey.ts
@@ -1,3 +1,0 @@
-//NOTE - auctionForm에서 사용하는 tansTack query 키
-export const AUCTION_FORM_QUERY_KEY = 'auctionForm';
-export const ADDRESS_ID_QUERY_KEY = 'addressId';
